requirements.txt
# Core ML / Training
torch>=2.1.0
transformers>=4.46.0
peft>=0.13.0
trl>=0.12.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
datasets>=3.0.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
scikit-learn>=1.5.0

# Config & CLI
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & Experiment Tracking
wandb>=0.18.0
tensorboard>=2.18.0

# Hugging Face Hub
huggingface-hub>=0.26.0

# Code Quality
ruff>=0.7.0
pytest>=8.0.0
